requirements.txt
opencv-python
facedetector-py
filetype
numpy
pytest
tf-keras
face_recognition
scikit-image
facenet-pytorch
deepface


